web-nfc/NFCErrorEvent_constructor.https.html

<!DOCTYPE html>
<title>NFCErrorEvent constructor</title>
<link rel="help" href="https://w3c.github.io/web-nfc/#dom-nfcerrorevent">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(() => {
    assert_equals(NFCErrorEvent.length, 2);
    assert_throws(new TypeError, () => new NFCErrorEvent('error'));
  }, 'NFCErrorEvent constructor without init dict');

  test(() => {
    const error = new DOMException;
    const event = new NFCErrorEvent('type', {error: error});
    assert_equals(event.type, 'type', 'type');
    assert_equals(event.error, error, 'error');
  }, 'NFCErrorEvent constructor with valid parameters');
</script>
